Record: use integers for seqnum, resp_seqnum, scid
I was avoiding this out of a sense of purity (the protocol treats them as
arbitrary 4-byte strings, any interretation is left up to the higher layers)
but in fact the one thing you do with sequence numbers is to compare them
with an inequality ("ignore all old messages", managed by comparing to a
high-water mark), and subchannel-ids are allocated sequentially (odd for one
side, even for the other). So it's easier to manage them as integers and let
encode_record/parse_record own the transformation into bytes.